feat: update path for theme-provider
