--- src/components/or-theme-provider/index.ts
@@ -1,3 +1,0 @@
-import OrThemeProvider from './or-theme.vue';
-
-export default OrThemeProvider;
