+++ src/components/theme-provider/index.ts
@@ -1,0 +1,4 @@
+import ThemeProvider from './theme.vue';
+
+export * from './use-theme'
+export default ThemeProvider;
