+++ src/components/theme-provider/use-theme.ts
@@ -1,4 +1,4 @@
-import { inject, ref, Ref } from "vue";
+import { computed, inject, reactive, ref, Ref } from "vue";
 import { ComponentOptions, DefaultTheme, injectionKey } from "@/types";
 import { merge } from 'lodash/fp';
 
@@ -9,8 +9,20 @@ type ReturnValue<T> = {
 
 export const useTheme = <T>(key?: keyof ComponentOptions): ReturnValue<T> => {
     const defaultTheme = ref<Partial<DefaultTheme>>({
-        primaryBg: '#000',
-        secondaryBg: 'teal',
+        primaryBg: '#1F2937',
+        secondaryBg: '#5f27cd',
+        dangerBg: '#e74c3c',
+        successBg: '#2ecc71',
+        infoBg: '#3498db',
+        text: {
+            primary: '#3c3c3f',
+            secondary: '#59595e'
+        },
+        grey: {
+            dark: '#cccccc',
+            dark2: '#e6ebee77',
+            dark3: '#e6ebee3b'
+        }
     });
 
     const componentTheme = ref<Partial<T>>();
@@ -18,8 +30,10 @@ export const useTheme = <T>(key?: keyof ComponentOptions): ReturnValue<T> => {
     const theme = (inject(injectionKey) as ComponentOptions);
 
     // theme may be undefined because the ThemeProvider isn't being used
-    if (theme && 'default' in theme) {
-        defaultTheme.value = merge(defaultTheme.value, theme.default);
+    if (theme) {
+        if ('default' in theme) {
+            defaultTheme.value = merge(defaultTheme.value, theme.default);
+        }
 
         if (key && key in theme) {
             componentTheme.value = (theme[key] as unknown) as Partial<T>;
@@ -33,7 +47,7 @@ export const useTheme = <T>(key?: keyof ComponentOptions): ReturnValue<T> => {
 }
 
 export const useStyles = <T extends {[key: string]: unknown}>(componentTheme: Ref<T | undefined>, defaultTheme: Ref<Partial<DefaultTheme>>) => {
-    return {
+    return computed(() => ({
         primaryBg: (componentTheme?.value?.primaryBg || defaultTheme.value.primaryBg),
-    }
+    }));
 }
